fix: use `url.parse` instead of the URL object
The docs for Node.js v6.x are wrong. In Node.js v6.11.1 (and about), the
`url` module exports an `Url` field, not `URL`. In Node.js v6.17.0 (and
about), `URL` seems to work as documented.

In both cases `url.parse` seems to work as intended, so I used that
instead, for compatibility.

https://nodejs.org/docs/latest-v6.x/api/url.html#url_constructor_new_url_input_base