feat: enable/disable escrow based on `device_config.json` field
If the field `billValidator.escrowEnabled` in `device_config.json` is
set and is a boolean, its value will be used to enable/disable the
escrow functionality of the bill validator. This new field is optional,
for backwards compatibility; if missing, the escrow functionality is not
explicitly enabled/disabled.

```json
"billValidator": {
  "deviceType": "genmega",
  "escrowEnabled": false,
  "rs232": {
    "device": "/dev/ttyS2"
  }
},
```